crm/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
